switching to different location, nothing important

+++ includes/utils.h
@@ -10,9 +10,9 @@
  * GNL utils.
  */
 int		get_next_line(int fd, char **line);
-char	*ft_strndup(char *src, int n);
+char	*ft_strndup(char *src, size_t n);
 char	*ft_strgjoin(char *s1, char *s2, int len_s2);
 size_t	ft_strlcpy(char *dest, const char *src, size_t size);
 size_t	ft_strlen(const char *s);
 
-#	endif
+#endif
